refactor parametric_analyser into folder with slightly less monstrously sized files in
